+++ include/NeoFOAM/dsl.hpp
@@ -6,5 +6,6 @@
 #include "dsl/explicit.hpp"
 #include "dsl/expression.hpp"
 #include "dsl/implicit.hpp"
-#include "dsl/operator.hpp"
+#include "dsl/spatialOperator.hpp"
+#include "dsl/temporalOperator.hpp"
 #include "dsl/solver.hpp"