+++ include/NeoFOAM/finiteVolume/cellCentred/operators/ddtOperator.hpp
@@ -6,9 +6,9 @@
 #include "NeoFOAM/fields/field.hpp"
 #include "NeoFOAM/core/executor/executor.hpp"
 #include "NeoFOAM/core/input.hpp"
-#include "NeoFOAM/dsl/operator.hpp"
+#include "NeoFOAM/dsl/spatialOperator.hpp"
 #include "NeoFOAM/mesh/unstructured.hpp"
-#include "NeoFOAM/finiteVolume/cellCentred/operators/fvccSparsityPattern.hpp"
+#include "NeoFOAM/finiteVolume/cellCentred/operators/sparsityPattern.hpp"
 
 namespace NeoFOAM::finiteVolume::cellCentred
 {
@@ -22,10 +22,14 @@ public:
 
     DdtOperator(dsl::Operator::Type termType, VolumeField<scalar>& field)
         : dsl::OperatorMixin<VolumeField<scalar>>(field.exec(), field, termType),
-          sparsityPattern_(SparsityPattern::readOrCreate(field.mesh())) {};
+          sparsityPattern_(SparsityPattern::readOrCreate(field.mesh())) {
 
-    void explicitOperation(Field<scalar>& source)
+          };
+
+    void explicitOperation(Field<scalar>& source, scalar t, scalar dt)
     {
+        const scalar rDeltat = 1 / dt;
+        const auto vol = getField().mesh().cellVolumes().span();
         auto operatorScaling = getCoefficient();
         auto [sourceSpan, field, oldField] =
             spans(source, field_.internalField(), oldTime(field_).internalField());
@@ -33,7 +37,7 @@ public:
             source.exec(),
             source.range(),
             KOKKOS_LAMBDA(const size_t celli) {
-                sourceSpan[celli] += field[celli] - oldField[celli];
+                sourceSpan[celli] += rDeltat * (field[celli] - oldField[celli]) * vol[celli];
             }
         );
     }
@@ -43,19 +47,24 @@ public:
         return sparsityPattern_->linearSystem();
     }
 
-    void implicitOperation(la::LinearSystem<scalar, localIdx>& ls)
+    void implicitOperation(la::LinearSystem<scalar, localIdx>& ls, scalar t, scalar dt)
     {
+        const scalar rDeltat = 1 / dt;
+        const auto vol = getField().mesh().cellVolumes().span();
         const auto operatorScaling = getCoefficient();
         const auto [diagOffs, oldField] =
             spans(sparsityPattern_->diagOffset(), oldTime(field_).internalField());
         const auto rowPtrs = ls.matrix().rowPtrs();
         const auto colIdxs = ls.matrix().colIdxs();
         std::span<scalar> values = ls.matrix().values();
+        std::span<scalar> rhs = ls.rhs().span();
         NeoFOAM::parallelFor(
             ls.exec(),
-            {0, coeff.size()},
+            {0, oldField.size()},
             KOKKOS_LAMBDA(const size_t celli) {
                 std::size_t idx = rowPtrs[celli] + diagOffs[celli];
+                values[idx] += operatorScaling[celli] * rDeltat * vol[celli];
+                rhs[celli] += operatorScaling[celli] * rDeltat * oldField[celli] * vol[celli];
             }
         );
     }
@@ -66,11 +75,10 @@ public:
         // do nothing
     }
 
-    std::string getName() const { return "DivOperator"; }
+    std::string getName() const { return "DdtOperator"; }
 
 private:
 
-    const VolumeField<scalar>& coefficients_;
     const std::shared_ptr<SparsityPattern> sparsityPattern_;
 };