+++ src/finiteVolume/cellCentred/operators/sparsityPattern.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: 2023 NeoFOAM authors
 
-#include "NeoFOAM/finiteVolume/cellCentred/operators/fvccSparsityPattern.hpp"
+#include "NeoFOAM/finiteVolume/cellCentred/operators/sparsityPattern.hpp"
 #include "NeoFOAM/fields/segmentedField.hpp"
 
 namespace NeoFOAM::finiteVolume::cellCentred